Extract FileSystem class to disinherit Repository from FileTree.
