+++ labs/demo-server/src/api/project/repo/fs.ts
@@ -41,7 +41,7 @@ export default apiController(function projectRepoFsApi(api, options) {
     },
     async function fsAdd({ body }, rep) {
       const res = await transaction((repo, out) =>
-        repo.add(body.to, "data" in body ? { data: body.data } : {}, out),
+        repo.fs.add(body.to, "data" in body ? { data: body.data } : {}, out),
       );
       rep.status(200).send(res);
     },
@@ -64,9 +64,9 @@ export default apiController(function projectRepoFsApi(api, options) {
     async function fsGetJsonData({ body }, rep) {
       const { repo } = projectService;
       const { from } = body;
-      const entry = repo.findPathEntry(from)!;
+      const entry = repo.fs.findPathEntry(from)!;
       // TODO: Respond with a 404 if (!entry)...
-      const data = repo.data(entry);
+      const data = repo.fs.data(entry);
       // CONSIDER: Should we compare client `ctime` to signal no-change here?
       rep.status(200).send({ id: entry.id, data });
     },
@@ -86,7 +86,7 @@ export default apiController(function projectRepoFsApi(api, options) {
     },
     async function fsMove({ body }, rep) {
       const res = await transaction((repo, out) =>
-        repo.move(body.from, body.to, out),
+        repo.fs.move(body.from, body.to, out),
       );
       rep.status(200).send(res);
     },
@@ -105,7 +105,9 @@ export default apiController(function projectRepoFsApi(api, options) {
       },
     },
     async function fsRemove({ body }, rep) {
-      const res = await transaction((repo, out) => repo.remove(body.from, out));
+      const res = await transaction((repo, out) =>
+        repo.fs.remove(body.from, out),
+      );
       rep.status(200).send(res);
     },
   );
@@ -124,8 +126,8 @@ export default apiController(function projectRepoFsApi(api, options) {
     },
     async function fsWrite({ body }, rep) {
       const res = await transaction(async (repo, out) => {
-        // await repo.write(body.to, (data) => {});
-        let entry = repo.findPathEntry(body.to);
+        // await repo.fs.write(body.to, (data) => {});
+        let entry = repo.fs.findPathEntry(body.to);
         if (!entry) {
           throw new Error(`Entry not found "${body.to}".`);
         }
@@ -133,7 +135,7 @@ export default apiController(function projectRepoFsApi(api, options) {
           if (!body.patches || !body.undo) {
             throw new Error(`Need data or patches to write to "${body.to}".`);
           }
-          entry = await repo.patch(
+          entry = await repo.fs.patch(
             entry,
             {
               ctime: body.ctime!,
@@ -146,7 +148,7 @@ export default apiController(function projectRepoFsApi(api, options) {
         } else if (typeof body.data === "undefined") {
           throw new Error(`Need data or patches to write to "${body.to}".`);
         } else {
-          entry = await repo.write(entry, body.data, out);
+          entry = await repo.fs.write(entry, body.data, out);
           return entry;
         }
       });
