+++ packages/core/src/FileTree.ts
@@ -36,16 +36,21 @@ export class FileTree {
   #rid: string;
   #nodes: FileTreeNodes;
   #root: Readonly<FileTreeRoot>;
-  #rootPath: string;
+  // #rootPath: string;
   #tx: number;
 
   constructor(options?: {
-    rootPath?: string;
+    // rootPath?: string;
     rid?: string;
     [INTERNAL]?: FileTreeInternal;
   }) {
-    const { rootPath, rid, [INTERNAL]: internal } = options ?? {};
-    this.#rootPath = rootPath ?? "";
+    const {
+      // rootPath,
+
+      rid,
+      [INTERNAL]: internal,
+    } = options ?? {};
+    // this.#rootPath = rootPath ?? "";
     this.#rid = rid ?? "";
     this.#nodes = internal?.nodes ?? new Map<string, Node>();
     this.#root = internal?.root ?? createRoot();
@@ -160,13 +165,13 @@ export class FileTree {
     return this.#root;
   }
 
-  get rootPath() {
-    return this.#rootPath;
-  }
+  // get rootPath() {
+  //   return this.#rootPath;
+  // }
 
-  protected set rootPath(value: string) {
-    this.#rootPath = value;
-  }
+  // protected set rootPath(value: string) {
+  //   this.#rootPath = value;
+  // }
   /** Current transaction number set by the driver. */
   get tx() {
     return this.#tx;
@@ -544,19 +549,19 @@ export class FileTree {
     }
     return this.getNodeDepth(node);
   }
-  /** Gets the full path to the given `entry` or the root path if no `entry`. */
-  fullPath(entry?: EntryOrId | null) {
-    if (!entry) return this.rootPath;
-    if (typeof entry === "string") {
-      const node = this.getEntry(entry);
-      if (!node) return this.rootPath;
-      const nodePath = this.path(node) ?? "";
-      return concatPath(this.rootPath, nodePath);
-    } else {
-      const nodePath = this.path(entry) ?? "";
-      return concatPath(this.rootPath, nodePath);
-    }
-  }
+  // /** Gets the full path to the given `entry` or the root path if no `entry`. */
+  // fullPath(entry?: EntryOrId | null) {
+  //   if (!entry) return this.rootPath;
+  //   if (typeof entry === "string") {
+  //     const node = this.getEntry(entry);
+  //     if (!node) return this.rootPath;
+  //     const nodePath = this.path(node) ?? "";
+  //     return concatPath(this.rootPath, nodePath);
+  //   } else {
+  //     const nodePath = this.path(entry) ?? "";
+  //     return concatPath(this.rootPath, nodePath);
+  //   }
+  // }
   /**
    * Returns the path for the given node id and `undefined` if id not found.
    */
@@ -801,9 +806,9 @@ function callEventHandlers<T, H extends EventHandler<T>>(
   }
 }
 
-function concatPath(a: string, b: string) {
-  if (a.endsWith("/")) {
-    return a + b;
-  }
-  return a + "/" + b;
-}
+// function concatPath(a: string, b: string) {
+//   if (a.endsWith("/")) {
+//     return a + b;
+//   }
+//   return a + "/" + b;
+// }
